feat(find-funders): show feedback card at end of search results
The thumbs up/down feedback was rendered inside NarrativeBlock,
immediately after the AI narrative paragraph and above the entity
list. Long result sets pushed the controls off-screen, so most
users never saw them and never rated a search.

Extract the feedback UI into a new SearchFeedback component and
render it at the bottom of each completed assistant turn (after
the entity list and any attachments) with a labeled "Helpful" /
"Not helpful" card, an explanatory prompt, and the existing
thumbs-down comment textarea. NarrativeBlock now only renders
the linkified narrative.

Submitted ratings still POST to /v2/agent/rating and Langfuse
scores them under the existing user_rating score name.
